multi task rep sharing

+++ tasks/sv_agreement.py
@@ -93,15 +93,15 @@ class WordSvAgreementVP(Task):
     return 2
 
   def get_loss_fn(self):
-    return ClassificationLoss(global_batch_size=tf.constant(self.task_params.batch_size), padding_symbol=tf.constant(-1))
+    return ClassificationLoss(global_batch_size=tf.constant(self.task_params.batch_size), padding_symbol=tf.constant(-1, dtype=tf.int64))
 
   def get_distill_loss_fn(self, distill_params):
-    return DistillLoss(tmp=tf.constant(distill_params.distill_temp), padding_symbol=tf.constant(-1))
+    return DistillLoss(tmp=tf.constant(distill_params.distill_temp), padding_symbol=tf.constant(-1, dtype=tf.int64))
 
   def get_probs_fn(self):
     return get_probs
 
 
   def metrics(self):
-    return [ClassificationLoss(global_batch_size=tf.constant(self.task_params.batch_size), padding_symbol=tf.constant(-1)),
+    return [ClassificationLoss(global_batch_size=tf.constant(self.task_params.batch_size), padding_symbol=tf.constant(-1, dtype=tf.int64)),
             tf.keras.metrics.SparseCategoricalAccuracy()]
